Project 1 - Assignment 1
1. Get the Calculator working as demonstrated in lectures 1 and 2. The
Autolayout portion at the end of the lecture is extra credit, but give
it a try because getting good at autolayout requires experience. If you
do not do autolayout, be sure to position everything so that it is
visible on all iPhones (i.e. the upper left corner of the scene).
2. Your calculator already works with floating point numbers (e.g. if
you touch 3 ↲ 4 ÷, it will properly show 0.75), however, there is no
way for the user to enter a floating point number directly. Fix this by
allowing legal floating point numbers to be entered (e.g. “192.168.0.1”
is not a legal floating point number!). You will have to add a new “.”
button to your Calculator. Don’t worry too much about precision or
significant digits in this assignment.
3. Add the following operations to your Calculator:
a. sin: calculates the sine of the top operand on the stack
b. cos: calculates the cosine of the top operand on the stack
c. π: calculates (well, conjures up) the value of π. For example, 3 π ×
should put three times the value of π into the display on your
calculator. Ditto 3 ↲ π x and also π 3 ×.
4. Add a UILabel to your UI which shows a history of every operand and
operation input by the user. Place it at an appropriate location in
your UI.
5. Add a C button that clears everything (your display, the new UILabel
you added above, etc.). The Calculator should be in the same state as
it is at application startup after you touch this new button.
6. Avoid the problems listed in the Evaluation section below. This list
grows as the quarter progresses, so be sure to check it again with each
assignment.